+++ apps/polpenaloza/package.json
@@ -14,11 +14,6 @@
   },
   "dependencies": {
     "@prisma/client": "^5.7.1",
-    "@tanstack/react-query": "^5.15.0",
-    "@trpc/client": "^10.45.0",
-    "@trpc/next": "^10.45.0",
-    "@trpc/react-query": "^10.45.0",
-    "@trpc/server": "^10.45.0",
     "@vercel/speed-insights": "^1.0.2",
     "@workspace/core-js": "workspace:^",
     "@workspace/ui": "workspace:^",
@@ -29,7 +24,6 @@
     "i18next-http-backend": "^2.4.1",
     "immer": "^10.0.3",
     "next": "^14.0.4",
-    "openai": "^4.24.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-i18next": "^13.5.0",
@@ -37,16 +31,11 @@
     "react-rotating-text": "^1.4.1",
     "react-text-transition": "^3.1.0",
     "sharp": "^0.32.6",
-    "sugar": "^2.0.6",
-    "superjson": "^2.2.1",
-    "tslib": "^2.6.2",
     "tsparticles": "^2.12.0",
     "tsparticles-engine": "^2.12.0",
-    "zod": "^3.22.4",
     "zustand": "^4.4.7"
   },
   "devDependencies": {
-    "@tanstack/react-query-devtools": "^5.15.0",
     "@types/debug": "^4.1.12",
     "@types/node": "^20.10.5",
     "@types/react": "^18.2.46",
@@ -71,6 +60,7 @@
     "dev:me": "next dev",
     "export": "next export",
     "format": "prettier --write .",
+    "generate:db": "prisma generate --accelerate",
     "lint": "next lint --fix",
     "start": "next start"
   },
