--- apps/polpenaloza/src/server/routers/movies.autocomplete.ts
@@ -1,27 +1,0 @@
-import { z } from 'zod'
-
-import { prisma } from '@/server/primsa'
-import { publicProcedure } from '@/server/trpc'
-
-export const moviesAutocomplete = publicProcedure
-  .input(
-    z.object({
-      search: z.string().nullish(),
-    })
-  )
-  .query(async ({ input }) => {
-    const { search } = input
-
-    const items = await prisma.movie.findMany({
-      take: 25,
-      where: {
-        name: {
-          contains: search || '',
-        },
-      },
-    })
-
-    return {
-      items,
-    }
-  })
