--- apps/polpenaloza/src/server/trpc.ts
@@ -1,39 +1,0 @@
-import { initTRPC } from '@trpc/server'
-import superjson from 'superjson'
-
-import { Context } from './context'
-
-const t = initTRPC.context<Context>().create({
-  /**
-   * @see https://trpc.io/docs/v10/data-transformers
-   */
-  transformer: superjson,
-  /**
-   * @see https://trpc.io/docs/v10/error-formatting
-   */
-  errorFormatter({ shape }) {
-    return shape
-  },
-})
-
-/**
- * Create a router
- * @see https://trpc.io/docs/v10/router
- */
-export const router = t.router
-
-/**
- * Create an unprotected procedure
- * @see https://trpc.io/docs/v10/procedures
- **/
-export const publicProcedure = t.procedure
-
-/**
- * @see https://trpc.io/docs/v10/middlewares
- */
-export const middleware = t.middleware
-
-/**
- * @see https://trpc.io/docs/v10/merging-routers
- */
-export const mergeRouters = t.mergeRouters
