+++ packages/ui/package.json
@@ -2,17 +2,12 @@
   "author": "@polpenaloza (https://github.com/polpenaloza)",
   "dependencies": {
     "@mdi/js": "^7.4.47",
-    "@tanstack/match-sorter-utils": "^8.8.4",
-    "@tanstack/react-table": "^8.11.2",
-    "@tanstack/react-virtual": "^3.0.1",
-    "@tanstack/table-core": "^8.11.2",
     "@workspace/core-js": "workspace:^",
     "classnames": "^2.5.0",
     "daisyui": "^3.9.4",
     "lodash": "^4.17.21",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-number-format": "^5.3.1",
     "usehooks-ts": "^2.9.1"
   },
   "devDependencies": {
